+++ Mods ChoGGi/Faster UI/Code/Script.lua
@@ -1,5 +1,7 @@
 -- See LICENSE for terms
 
+local pcall = pcall
+
 local mod_EnableMod
 
 -- Fake funcs
@@ -7,7 +9,7 @@ local ChoOrig_GetPreciseTicks = GetPreciseTicks
 local function ChoFake_GetPreciseTicks()
 	return 0
 end
---
+-- Faster pins
 local ChoOrig_PinsDlg_SetVisible = PinsDlg.SetVisible
 function PinsDlg:SetVisible(visible, instant, ...)
 	if mod_EnableMod then
@@ -19,7 +21,7 @@ function PinsDlg:SetVisible(visible, instant, ...)
 	pcall(ChoOrig_PinsDlg_SetVisible, self, visible, instant, ...)
 	GetPreciseTicks = ChoOrig_GetPreciseTicks
 end
---
+-- Open build menu
 local ChoOrig_XBuildMenu_EaseInButton = XBuildMenu.EaseInButton
 function XBuildMenu:EaseInButton(button, start_time, ...)
 	if mod_EnableMod then
@@ -27,7 +29,7 @@ function XBuildMenu:EaseInButton(button, start_time, ...)
 	end
 	return ChoOrig_XBuildMenu_EaseInButton(self, button, start_time, ...)
 end
---
+-- Switch to map overview
 local ChoOrig_OverviewModeDialog_GetCameraTransitionTime = OverviewModeDialog.GetCameraTransitionTime
 function OverviewModeDialog.GetCameraTransitionTime(...)
 	if mod_EnableMod then
@@ -35,6 +37,8 @@ function OverviewModeDialog.GetCameraTransitionTime(...)
 	end
 	return ChoOrig_OverviewModeDialog_GetCameraTransitionTime(...)
 end
+
+-- Fade to black for map switch buttons
 
 -- pp is too soon for mod options, so we default to "enabled" for it
 function OnMsg.ClassesPostprocess()
@@ -56,6 +60,7 @@ local function UpdateFade(time)
 	template.FadeOutTime = time
 end
 
+-- Update mod options
 local function ModOptions(id)
 	-- id is from ApplyModOptions
 	if id and id ~= CurrentModId then