+++ Mods ChoGGi/Toggle Visible Vegetation/metadata.lua
@@ -4,17 +4,18 @@ return PlaceObj("ModDef", {
 	"lua_revision", 1007000, -- Picard
 	"steam_id", "2162375075",
 	"pops_any_uuid", "e42779d0-c633-4a69-b3c4-69db21c0216f",
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
-	"image", "Preview.png",
+	"version_minor", 2,
+	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua",
 	},
 	"has_options", true,
 	"TagCosmetics", true,
-	"description", [[Adds mod option to hide trees.
+	"description", [[
+Adds mod option to hide trees.
 Includes mod option to also toggle bushes (needs to be enabled if turning visibility back on).